% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_4}
\alias{col_4}
\title{Columns wrappers}
\usage{
col_4(...)
}
\arguments{
\item{...}{you ui components}
}
\description{
These are convenient wrappers around
\code{column(12, ...)}, \code{column(6, ...)}, \code{column(4, ...)}...
}
