% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_spring.R
\name{create_module}
\alias{create_module}
\title{Create a baseline Module for your shinyapp}
\usage{
create_module(mod_name)
}
\arguments{
\item{mod_name}{The Modules Name}
}
\description{
Create a baseline Module for your shinyapp
}
