% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_spring.R
\name{create_shinyapp}
\alias{create_shinyapp}
\title{Creates the app.R using the templates defined in inst/cheetah/*}
\usage{
create_shinyapp(params, target_file = "app.R", template_file = NULL)
}
\arguments{
\item{params}{the params for creating app.R}

\item{target_file}{(optional) default is app.R , you can change this}

\item{template_file}{(optiona) this is an override.Function expect that you params$template_file is set in params when template_file is NULL}
}
\description{
Based on the configuration defined in config.yml the app.R code is created
}
