% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_spring.R
\name{create_targets}
\alias{create_targets}
\title{Create _targets file based on shinyspring approach}
\usage{
create_targets(params, filename = "_targets.R", overwrite = FALSE)
}
\arguments{
\item{params}{the params}

\item{filename}{(optional) defaults to _targets.R}

\item{overwrite}{existing _targets.R file is already present. (optional) defaults to FALSE}
}
\description{
Create _targets file based on shinyspring approach
}
