% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_manager.R
\name{ds_parameters}
\alias{ds_parameters}
\title{Fetch ds params from config}
\usage{
ds_parameters(controller, ds_name)
}
\arguments{
\item{controller}{the  app_master}

\item{ds_name}{the name of DS}
}
\description{
Fetch ds params from config
}
