% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_manager.R
\name{ds_validate_and_prep}
\alias{ds_validate_and_prep}
\title{Validate Datasets based on rules and prep conduct basic data prep
Typically this method is called from _targets or at Design time when building the app
Not ideal to call this method from shiny app.
Note: validtion is done only if the dataset has the following set in config file
ds_info_type and ds_info_url
supported ds_info_type : google , excel and csv}
\usage{
ds_validate_and_prep(controller, ds_names)
}
\arguments{
\item{controller}{app_master}

\item{ds_names}{the dataset names}
}
\value{
logical FALSE if data validation Warnings or Errors elase TRUE
}
\description{
Validate Datasets based on rules and prep conduct basic data prep
Typically this method is called from _targets or at Design time when building the app
Not ideal to call this method from shiny app.
Note: validtion is done only if the dataset has the following set in config file
ds_info_type and ds_info_url
supported ds_info_type : google , excel and csv
}
