% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_str}
\alias{is_valid_str}
\title{checks if valid string obj}
\usage{
is_valid_str(str)

is_valid_str(str)
}
\arguments{
\item{str}{string}
}
\value{
TRUE if valid FALSE when NULL , NA or length == 0

TRUE/FALSE
}
\description{
checks if valid string obj

Check if valid string
}
