% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_list_to_str}
\alias{merge_list_to_str}
\title{Converts List of str to unique str seperated by sep}
\usage{
merge_list_to_str(str_list, sep = ", ")
}
\arguments{
\item{str_list}{the lis}

\item{sep}{(optional) default is ,}
}
\value{
string
}
\description{
Converts List of str to unique str seperated by sep
}
