% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_config.R
\name{mod_definition}
\alias{mod_definition}
\title{Print props needed for a mod}
\usage{
mod_definition(mod_ref)
}
\arguments{
\item{mod_ref}{the mod_ref name : eg: rmd_mod}
}
\description{
prints the mod properties and default values
}
