% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_row}
\alias{new_row}
\title{creates the standardized row for app_master mdata}
\usage{
new_row(sr_num, ds, ds_name, ds_params)
}
\arguments{
\item{sr_num}{the serial number for the row. Index}

\item{ds}{the dataset}

\item{ds_name}{the unique name for the dataset you wand to refer it with to access it in the application}

\item{ds_params}{the config file params for file}
}
\value{
row the row
}
\description{
create a row that goes into mdata of reactiveValues(rvals)
}
