% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_targets.R
\name{tar_shinyspring}
\alias{tar_shinyspring}
\title{ease of use method for initializing shinyspring in a target}
\usage{
tar_shinyspring(config_file)
}
\arguments{
\item{config_file}{the config.yml for shinyspring application.}
}
\value{
master list with controller , registry and ds_names
}
\description{
ease of use method for initializing shinyspring in a target
}
\examples{
# include the following 2 targets in your _targets.R file
}
