% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_spring.R
\name{test_config_file}
\alias{test_config_file}
\title{Checks if the config file is good to go
Note : this runs a bunch of testthat tests}
\usage{
test_config_file(params)
}
\arguments{
\item{params}{all params}
}
\description{
Checks if the config file is good to go
Note : this runs a bunch of testthat tests
}
